Se crea la capa dto, y se finaliza la insercion en ambas tablas

+++ src/main/java/com/cotizador/cotizador_seguros_api/repositorios/IUsuarioRepositorio.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IUsuarioRepositorio extends JpaRepository<Usuario,Long>{
+public interface IUsuarioRepositorio extends JpaRepository<Usuario,Integer>{
 
 }
