+++ src/main/java/com/cotizador/cotizador_seguros_api/repositorios/IVehiculoRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IVehiculoRepositorio extends JpaRepository<Vehiculo, Long> {
+public interface IVehiculoRepositorio extends JpaRepository<Vehiculo, Integer> {
 }
